[#4] Fix error for invalid prefix path when using regex expressions
The current nginx ingress version sets the default for strict-validate-path-type to true, which only allows alphanumeric for the path. For Nexus, we are using a regex expression, which is stated as invalid. As the ingress is invalid, the docker registry is not accessible.

By setting strict-validate-path-type to false the validation is disabled for the moment.